Gtk-3.22&4.0: Add 'fallback-mode' widget styling for non-native Gtk+ backends
This commit is the big change in some non-native apps, like damned
LibreOffice and Webkit2Gtk. This brings 'Gtk+ 2' like flat elements
in GtkButton, GtkEntry and GtkComboBox.

* Removed any drop-shadows from buttons, combos and entries.
* Draw solid thin borders around those widgets.
* Add hover/active state background colours like Gtk+ 2 theming.

So those widgets are rendered as 'non-elevation' elements though,
but anyway this commit is really useless on native Gtk+ 3 apps and
mozilla's apps...