Gtk-3.22&4.0: [Gnome-control-center] Draw fake-trough behind the arrow-pointer in Night-Light widget
That arrow-pointer and other elements are rendered by its own code-base
(cairo?), meant that widget ignores Gtk+ colour variant at all.
Once that widget was rendered with dark-variant, it's hard to see
the arrow-pointer what pointed to the current position of sunlight.

this is really bad hack by drawing the trough on the parent GtkBox,
so in future, this should be dropped/replaced...