Gtk-3.20&3.22&4.0: [Budgie-desktop] Add Alt-Tab apps switcher styling (2)
In Gtk+ 3.22 or higher, The width/height of GtkWindow strongly depends
on its child element allocations. And Adapta is the scaling Gtk+
theme. Once this combo happnes, we can not keep the width defined
by upstream's ui file(s) like 'max_chars' and/or 'max_children_per_line'.

So let's add the blocker for minimum width of GtkLabel via non-scaled
px value, at this time, 13.33px x 45 chars = 600px was applied. Then we
can keep at least the number of GtkFlowBoxChilds (8) in a sigle row
even if we've set 0.9 or 0.8 to font-scaling-factor...