Gtk-3.x/4.0: Refactor 'fallback-mode' buttons/entries design
* Employ new 'light_color' for dark-variant.
* Re-define insensitive background-colors.
* Rewrite toggle images.
* No longer used base_color for those widgets, use light_color instead.

[Side-effect]
* In LibreOffice's retractable text input area stacks with only just
'.view' style-class without textview node. Means internal region
of that input-box can not be drawn with light_colors but base_color.
This is the upstream issue. Maybe their poor 'multi layered elements'
might not be able to blend with transparent background, so they seem
to put .view style-class inside the entry node. If so, I don't have
any ideas why they ignored the textview node.
If textview node was existed in that place, we could override that
input-box background colour... :/