Gtk-3.20&3.22&4.0: [Budgie-desktop] Trick damned nm-applet NaChild
Add negative margin to the parent NaTray box container to trick the
nm-applet NaChild. Basically, 36px or smaller 'size' was set to
Budgie-panel, NaTray's allocation seems to be shrunk like '24x16' or
so, once that happned, NaChilds can't hang over that allocation.
It means 16px un-recolourable icon images are used in panel...

[Side-effect]
On my rough tests, this commit introduces the unwanted shrinking
of clickable area of NaChilds, it's so hard to do 'edge to click'.