Gtk-3.20&3.22&4.0: [Chrome(ium)] Follow the upstream nav_button_provider_gtk3.cc changes
That code affects when we disabled "Use system title bar and borders".
And upstream did not check the '.close', '.maximize'... classes yet,
so we need to re-define button sizing as '.titlebutton' globaly...

The biggest problem for us is that code searches button icons from
icon-theme directly and ignores gtk-theme's implementations at all.