Gtk-3.22&4.0: [GtkToolBar] Rewrite child sizing and styling
GtkToolBar is one of the damned widgets when we'd like to estimate
the child spacing and sizing especially in Nemo, Caja, Glade and some
other classic stacks. It brought some unwanted, broken
circular image-buttons in some apps. Because some buttons use 16px
symbolic SVGs, but others sometimes draw 24px PNGs or so...

 * Share the same button/entry sizing in all variants.
 * Kill vertical padding values of childs.
 * Stop using circular image-buttons in generic toolbar node.