Optimize by utilizing `bash`, and condense code
**The following is taken from my commit message on the below link:**
https://github.com/terminalforlife/adapta-midnight/commit/5e996bd24ce203ff87995bbf62e9f71a7b3f5adf

Since `bash` is being used, it seems logical to _actually_ use as many
such features as possible and appropriate, of which there are plenty.

A fair bit was done here, so I'll try to be concise:

  * Three functions were created for condensed code and easier editing.

  * The fetching and/or assignment of many variables were optimized.

  * Additional variables and strings were correctly quoted.

  * An unprotected string causing undesired results was corrected.

Overall, I believe this script is now a _lot_ easier to read and edit.